isolate the fake lexer
